Settings are loaded from classpath, candidate entities can be restricted.
IMPORTANT: The update requires additional tables in the database.

You can get the tables like this:
curl -O
http://www.mpi-inf.mpg.de/yago-naga/aida/download/entity-repository/AIDA
_entity_repository_2010-08-17_UPDATE.sql.bz2
bzcat AIDA_entity_repository_2010-08-17_UPDATE.sql.bz2 > psql
<YOUR_EXISTING_AIDA_DB>

[Changes]

- Settings and stopwords are now loaded from the classpath. Make sure
to have all the properties files in the root of your classpath and the
stopword tokens in the tokens/ folder on your classpath if you are not
using our main jar.
- The number of candidates can be restricted by a global ranking. Set
the maxEntityRank in the DisambiguationSettings, e.g. to 0.1 to only
disambiguate to the top 10% of all entities in AIDA (sorted by the
number of incoming links in Wikipedia, the more, the higher)
- Additional helper classes